+++ my-app/src/App.jsx
@@ -5,6 +5,7 @@ import Herosection from "./components/Herosection.jsx";
 import Logo from "./components/Logo.jsx";
 import Services from "./components/Services.jsx";
 import About from "../src/components/About.jsx";
+import Team from '../src/components/Team.jsx';
 
 function App() {
   return (
@@ -14,6 +15,7 @@ function App() {
       <Logo />
       <Services />
       <About />
+      <Team/>
     </>
   );
 }
