--- my-app/src/Js/About.js
@@ -1,17 +1,0 @@
-const about={
-    titleAbout:" What Make Us Different?",
-    descriptionAbout:"Check out our best service you can possibly orders in building your company and don't forget to ask via our email or our customer service if you are interested in using our services ",
-    img:[
-        {id:1,url:"../src/assets/Vector1.png"},
-         {id:2,url:"../src/assets/Vector2.png"},
-          {id:3,url:"../src/assets/Vector3.png"},
-           {id:4,url:"../src/assets/Vector5.png"}
-    ],
-    logotitle:[
-        {id:5,logotitl:"Experienced"},
-        {id:6,logotitl:"competitive price"},
-        {id:7,logotitl:"On Time"},
-        {id:8,logotitl:"Best Materials"},
-    ]
-}
-export default about;
