--- my-app/src/Js/Herosection.js
@@ -1,19 +1,0 @@
-const herosection={
-    title1:"We Provide",
-    title2:"Architectural design",
-    title3:"and Construction ",
-    description:"More than 100 building and housing projects that we have built. The building owner chose us over othe​r contractors in Jakarta, because our work is different",
-   herosectionButton:"Discover more",
-   Number1:"300",
-   Number2:"900",
-   Number3:"20",
-   sign:"+",
-   text1:"Happy Client",
-   text2:"Amazing projects",
-   text3:"Awards Winning",
-   imgdivdecs1:"General Project ",
-   imgdivdecs2:"25 Years Operated  ",
-   imgdivdecs3:"As a trusted general project that has been operating for 25 years, our commitment is always to prioritize our client satisfaction Operated",
-}
-
-    export default herosection;
