+++ my-app/src/Json/AboutCardData.js
@@ -1,0 +1,27 @@
+const aboutCardData = [
+  {
+    id: 1,
+    img: "../src/assets/Vector1.png",
+    title: "Experienced",
+    desc: "Our experience of 25 years of building and making achievements in the world of development",
+  },
+  {
+    id: 2,
+    img: "../src/assets/Vector2.png",
+    title: "competitive price",
+    desc: "The prices we offer you are very competitive without reducing the quality of the company's work in the slightest",
+  },
+  {
+    id: 3,
+    img: "../src/assets/Vector3.png",
+    title: "On Time",
+    desc: "We prioritize the quality of our work and finish it on time",
+  },
+  {
+    id: 4,
+    img: "../src/assets/Vector5.png",
+    title: "Best Materials",
+    desc: "The material determines the building itself so we recommend that you use the best & quality materials in its class.",
+  },
+];
+export default aboutCardData;
