+++ my-app/src/Json/Logo.js
@@ -1,0 +1,8 @@
+const images = [
+  { id: 1, url: "../src/assets/logo1.png" },
+  { id: 2, url: "../src/assets/logo2.png" },
+  { id: 3, url: "../src/assets/logo3.png" },
+  { id: 4, url: "../src/assets/logo4.png" },
+  { id: 5, url: "../src/assets/logo5.png" },
+];
+export default images;
