+++ my-app/src/Json/Servicess.js
@@ -1,0 +1,34 @@
+const servicess = {
+  title: "Our Excellent Services",
+  description:
+    "Check out our best service you can possibly orders in building your company and don't forget to ask via our email or our customer service if you are interested in using our services",
+  img: [
+    { id: 1, url: "../src/assets/services1.jpg" },
+    { id: 2, url: "../src/assets/services2.jpg" },
+    { id: 3, url: "../src/assets/services3.jpg" },
+  ],
+  imgtitle: [
+    { id: 4, imgtitle: "Industrial" },
+    { id: 5, imgtitle: "Commercial" },
+    { id: 6, imgtitle: "Resedential" },
+  ],
+  imgdesc: [
+    {
+      id: 7,
+      imgdesc:
+        "Industrial development is our main line of business. We do Factory Construction, Warehouse and others",
+    },
+    {
+      id: 8,
+      imgdesc:
+        "Our experience building in the Commercial field includes Showrooms,Supermalls and Office Buildings",
+    },
+    {
+      id: 9,
+      imgdesc:
+        "Residential development is the beginning that has shaped us to this day. Our development includes Houses & Apartments",
+    },
+  ],
+};
+
+export default servicess;
