+++ my-app/src/Json/Team.js
@@ -1,0 +1,14 @@
+const team={
+    img:"../src/assets/team.jpg",
+    person:{
+        img:"../src/assets/person.jpg",
+        title:"Dianne Russell",
+        desc:"More than 20 years of experience in the field architecture and has worked on project up to 100+"
+    },
+    title:"Meet and talk with our best architecture",
+    desc:"All our teams are professional and competent in their fields and will help you realize your dream building with the excellent result.",
+    teamBtn:"See all team",
+    footertext:"How it works ",
+    sign:"https://cdn-icons-png.flaticon.com/128/15714/15714609.png"
+}
+export default team;
