+++ my-app/src/components/About.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import about from "../Js/About";
-import aboutCardData from "../Js/AboutCardData";
+import about from "../Json/About";
+import aboutCardData from "../Json/AboutCardData";
 import Card from "./Card";
 
 const About = () => {
