+++ my-app/src/components/Button.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import herosection from '../Js/Herosection';
-const Button = () => {
+
+const Button = ({children}) => {
     return (
         <>
             <button className='h-14 bg-linear-to-r from-cyan-500 to-blue-500 text-white ms-20 w-50'>
-                {herosection.herosectionButton}
+                {children}
             </button>
         </>
     )
